Make all-layers sentinel cast explicit in CvPlot::removeUnit
plotManager().RemoveUnit() takes uint uiLayerID; passing int literal -1
caused an implicit sign-change (int(-1) to uint(4294967295)).

The sentinel value UINT_MAX means remove from all layers -- this is the
designed contract, as RemoveUnit itself guards with:
  if (uiLayerID != static_cast<uint>(-1))
CvPlot::removeUnit uses all-layers because its three call sites include
two paths (desync cleanup, DLL interface) where the unit layer is
unknown, making the sweep the only safe choice.

Use static_cast<uint>(-1) to match the implementation's own guard
expression. No behaviour change.

+++ CvGameCoreDLL_Expansion2/CvPlot.cpp
@@ -13418,7 +13418,7 @@ void CvPlot::removeUnit(CvUnit* pUnit, bool bUpdate)
 		gGlobals.getDLLIFace()->sendChat(msg, CHATTARGET_ALL, NO_PLAYER);
 	}
 
-	GC.getMap().plotManager().RemoveUnit(pUnit->GetIDInfo(), m_iX, m_iY, -1);
+	GC.getMap().plotManager().RemoveUnit(pUnit->GetIDInfo(), m_iX, m_iY, static_cast<uint>(-1));
 
 	if(bUpdate)
 	{
